feat: refactor movie details with hook
