+++ src/components/movie-item/use-movie-detail.ts
@@ -1,0 +1,17 @@
+import {useMovie} from '@hooks/use-movies'
+import {useParams, useSearchParams} from 'react-router-dom'
+
+export function useMovieDetails() {
+  // Get the id from the route params or query parameters
+  const {id} = useParams<{id: string}>()
+  const [searchParams] = useSearchParams()
+  const movieName = searchParams.get('name')
+
+  const identifier =
+    movieName ?? (id && !isNaN(Number(id)) ? parseInt(id, 10) : null)
+  if (!identifier) return {movie: null, isLoading: false, hasIdentifier: false}
+
+  const {data: movie, isLoading} = useMovie(identifier)
+
+  return {movie, isLoading, hasIdentifier: true}
+}
